Allow fields to define a `display_value` method
The admin utility method `display_for_field`, which is called to render
values in the changelist and for readonly fields in the form, describes
behaviour for built-in fields, but there is no way to extend this for
custom fields. This change introduces a `display_value` hook which will
be called if the field defines it to render the value.