Use relation attributes when creating formfields
Instead of checking for instances of ManyToManyField and ForeignKey in
the admin's formfield_for_dbfield method, use the relation attributes from
the field's Meta class.

Note that even though the original code only checked for those two
classes, it implicitly included OneToOneField as well since that is a
subclass of ForeignKey, so we need to check field.one_to_one as well as
field.many_to_one.